pkg/sensors : Move strings.Split(arg.Resole, ".") in dedicated func
Because we want to include argument filtering in the resolve path, we
need to make a split on the "." as long as one on the "[" when it
matches a bracket. This commit add a function like this, as long the
test.
One other thing is that it changes the way the "padding" in the resolve
path is registered. Until now, nested ptr such as (struct cred**) were
registered as "" in the btf path. Now, we consider the same as getting
the first parameter of an array. So "" is replaced by "[0]". But the
behaviour remain the same.

Signed-off-by: Tristan d'Audibert <tdaudibe@isovalent.com>