pkg/btf : Change `addPaddingOnNestedPtr` to set [0] instead of ""
The idea of this commit is to consider nested pointers as a dynamic
array. Currently, for each of those ptrs, an entry is added in the
`api.ConfigBTFArg` array. By replacing it as [0], we give them the same
behavior as resolving an array.

This commit also include the static test for `parseArrayIdxStr`.

Signed-off-by: Tristan d'Audibert <tdaudibe@isovalent.com>