pkg/btf : Move MaxBTFArgDepth check in btf for consistency
The `ResolveBTFPath` should be self-sufficient in term of limitation
validation.
Until now the function could have crashed if the given `pathToFound` was
longer than `api.MaxBTFArgDepth`.

Furthermore, introducing `ErrMaxBTFDepth`. This error is different than the
others because it's not directly related to BTF. So because of this
separate error, we can distinguish it from the others.

Signed-off-by: Tristan d'Audibert <tdaudibe@isovalent.com>